EventUtils Pega usability review: fix Hwnd truncation, JSON adapters, enums
Resolves all six recommended changes from the EventUtils Pega API usability
review:
- Fixed a real correctness bug: EventData.Hwnd was a 32-bit uint, truncating
  a 64-bit window handle via unchecked((uint)hwnd.ToInt64()) in
  WinEventEngine.Enrich. It's now a 64-bit long capturing the handle without
  loss, propagated through EventFilter's button-child cache/lookup and
  ThrottleDebounce's per-handle dedupe key, both of which were also keyed by
  the truncated uint before this pass.
- Converted EventData's public fields to properties ({ get; internal set; }),
  removing IncludeFields from its JSON serializer options accordingly.
- Added a JSON+IntPtr-handle companion overload for GetNextEvent and all 8
  WaitForX methods (9 methods total), sharing their existing private cores.
- Added GetNextEventsJson for batch subscription consumption, matching
  DumpRecentEvents' existing JSON-array shape.
- Added BuildFilterJson, building the filter JSON from scalar parameters
  (process, processesCsv, className, titleContains, titleMatches,
  hasButtonChildren, excludeSelf) instead of requiring hand-authored JSON;
  processesCsv is the scalar alternative to EventFilter.AnyOfProcesses.
- Added EventName (20 values) and EventOverflowPolicy (DropOldest,
  DropNewest, Block) enums, with enum overloads of SetDebounce/
  SetQueueLimits; the original string-based overloads remain.

Added tests for the Hwnd fix (a regression test proving ThrottleDebounce no
longer collapses two handles that differ only in their high 32 bits),
BuildFilterJson, the enum overloads, and guard behavior for the new JSON
overloads. All 57 tests pass (46 existing + 11 new) - this test project runs
on Linux (no UseWindowsForms/UseWPF).

+++ src/eventutils/EventFilter.cs
@@ -215,11 +215,11 @@ namespace EventAutomation
         // per hwnd for a short window so repeated events for the same dialog do
         // not re-enumerate; a hung window can at worst stall the pump once per
         // cache entry, not once per event.
-        private static readonly ConcurrentDictionary<uint, (long Ticks, bool Found)> ButtonChildCache =
-            new ConcurrentDictionary<uint, (long, bool)>();
+        private static readonly ConcurrentDictionary<long, (long Ticks, bool Found)> ButtonChildCache =
+            new ConcurrentDictionary<long, (long, bool)>();
         private static readonly TimeSpan ButtonChildCacheTtl = TimeSpan.FromMilliseconds(500);
 
-        private static bool HasButtonChild(uint hwnd)
+        private static bool HasButtonChild(long hwnd)
         {
             if (hwnd == 0)
                 return false;
@@ -228,7 +228,7 @@ namespace EventAutomation
                 return cached.Found;
             bool found = false;
             WinEventInterop.EnumChildWindows(
-                new IntPtr((long)hwnd),
+                new IntPtr(hwnd),
                 (child, lParam) =>
                 {
                     var sb = new System.Text.StringBuilder(64);
